refactor: remove unused package, and sort packages
