+++ cmd/knit/subcommands/common/testdata/current/.gitignore
